==> 0_data_carmax.ipynb <==
# CarMax Analytics Showcase
- Competition: https://analyticsshowcase.carmax.com
- Data: https://github.com/kmx-analytics-showcase/Winter-2023

## Data Load and Describe

# load packages
import numpy as np
import pandas as pd
from pandas.api.types import CategoricalDtype

# load data
file = 'data/data_winter2023.csv'
df = pd.read_csv(file
                ,dtype={'price':'category'
                        ,'appraisal_offer':'category'
                        ,'market':'category'

                        ,'mileage':'category'
                        #,'make':'category'
                        #,'body':'category'
                        ,'color':'category'

                        ,'mileage_appraisal':'category'
                        #,'make_appraisal':'category'
                        #,'body_appraisal':'category'
                        ,'color_appraisal':'category'
                }
                )
df.shape

# organize columns

# rename some fields
df.rename(columns={'trim_descrip':'premium_trim'
                  ,'trim_descrip_appraisal':'premium_trim_appraisal'
                  ,'appraisal_offer':'price_appraisal'
                  ,'engine':'engine_disp'
                  ,'engine_appraisal':'engine_disp_appraisal'
                  }, inplace=True)

# move market var column location in df
market = df.pop('market')
df.insert(0, 'market', market)

# move price_appraisal var column location in df
price_appraisal = df.pop('price_appraisal')
df.insert(df.columns.get_loc('model_year_appraisal'), 'price_appraisal', price_appraisal)

# move online_appraisal_flag var column location in df
online_appraisal_flag = df.pop('online_appraisal_flag')
df.insert(df.columns.get_loc('price_appraisal'), 'online_appraisal_flag', online_appraisal_flag)

df.info()

# make prices categorical with correct order
appraisal_cat = CategoricalDtype(
    ['$0k to $5k', '$5k to $10k', '$10k to $15k', '$15k to $20k', 
    '$20k to $25k', '$25k to $30k', '$30k to $35k', '$35k to $40k', 
    '$40k+'], 
    ordered=True
)
price_cat = CategoricalDtype(
    ['$0 to $15k', '$15k to $20k', '$20k to $25k', '$25k to $30k',
       '$30k to $35k', '$35k to $40k', '$40k to $45k', '$45k to $50k',
       '$50k to $55k', '$55k to $60k', '$60k to $65k', '$65k to $70k',
       '$70k+'], 
    ordered=True
)
df['price_appraisal'] = df['price_appraisal'].astype(appraisal_cat)
df['price'] = df['price'].astype(price_cat)

# make mileage categorical with correct order
mileage_app_cat = CategoricalDtype(
    ['0 to 10k miles', '10k to 20k miles', '20k to 30k miles',
     '30k to 40k miles', '40k to 50k miles','50k to 60k miles',
     '60k to 70k miles', '70k to 80k miles', '80k to 90k miles',
     '90k to 100k miles','100k to 110k miles', '110k to 120k miles',
     '120k to 130k miles', '130k to 140k miles','140k to 150k miles',
     '150k to 160k miles', '160k to 170k miles', '170k to 180k miles',
    '180k to 190k miles', '190k to 200k miles', '200k+ miles'],
    ordered=True
)
mileage_cat = CategoricalDtype(
   ['0 to 5k miles', '5k to 10k miles', '10k to 15k miles',
    '15k to 20k miles', '20k to 25k miles', '25k to 30k miles',
    '30k to 35k miles', '35k to 40k miles', '40k to 45k miles',
    '45k to 50k miles', '50k to 55k miles', '55k to 60k miles',
     '60k to 65k miles', '65k to 70k miles', '70k to 75k miles',
     '75k to 80k miles', '80k to 85k miles', '85k to 90k miles',
     '90k to 95k miles', '95k to 100k miles', '100k+ miles'],
    ordered=True
)
df['mileage_appraisal'] = df['mileage_appraisal'].astype(mileage_app_cat)
df['mileage'] = df['mileage'].astype(mileage_cat)

# make engine displacement numeric
df['engine_disp'] = df['engine_disp'].str.slice(0,3).astype('float')
df['engine_disp_appraisal'] = df['engine_disp_appraisal'].str.slice(0,3).astype('float')

df.info(show_counts=False)

df.nunique()

# numeric
df.describe().round(decimals = 1).T

# non-numeric
df.describe(include = ['object','category']).T

# data sample
df.sample(5).T

## Missingness

- Analyze missing values (either NA or filler)

### NA's

# na counts
pd.DataFrame({'count':df.isna().sum()
             ,'percent':(df.isna().sum()/df.shape[0]).round(decimals=2)}
    ).sort_values(by='count', ascending=False)

### 0's

# 0 counts in numeric vars
(df[['model_year','engine_disp','cylinders','mpg_city','mpg_highway','horsepower','fuel_capacity'
    ,'model_year_appraisal','engine_disp_appraisal','cylinders_appraisal','mpg_city_appraisal','mpg_highway_appraisal','horsepower_appraisal','fuel_capacity_appraisal']] < 0.1).sum()

## Imputation

#### Trim
- 16% of appraisals and 12% of purchases were missing trim
- Imputation: assumed not premium

# create imputed flags
df['imputed_flag_premium_trim'] = np.where(df['premium_trim'].isna(), 1, 0)
df['imputed_flag_premium_trim_appraisal'] = np.where(df['premium_trim_appraisal'].isna(), 1, 0)

# recode trim
# assumes na is not premium
df['premium_trim'] = np.where(df['premium_trim'] == 'Premium', 1, 0)
df['premium_trim_appraisal'] = np.where(df['premium_trim_appraisal'] == 'Premium', 1, 0)

#### Model
- 11% of appraisals are missing model
- Imputation: assume the most common model for each make

# create imputed flag
df['imputed_flag_model_appraisal'] = np.where(df['model_appraisal'].isna(), 1, 0)

# mode of model by make
# TODO: improve to mode by make/year
model_modes = df.groupby(['make_appraisal'])['model_appraisal'].agg(pd.Series.mode) # compute modes by make
na_indexer = df['model_appraisal'].isna() # null obs
df.loc[na_indexer,'model_appraisal'] = df.loc[na_indexer].apply(lambda x: model_modes.loc[x['make_appraisal']], axis=1) # apply modes by make
df.loc[df['model_appraisal'].map(type) == np.ndarray,'model_appraisal'] = np.nan # without mode set to null

#### Body
- 5% of body is missing
- Imputation: assume mode of body for each make

# create imputed flag
df['imputed_flag_body'] = np.where(df['body'].isna(), 1, 0)

# mode of body by make
# TODO: improve to mode by make/model
body_modes = df.groupby(['make'])['body'].agg(pd.Series.mode) # compute modes by make
na_indexer = df['body'].isna() # null obs
df.loc[na_indexer,'body'] = df.loc[na_indexer].apply(lambda x: body_modes.loc[x['make']], axis=1) # apply modes by make
df.loc[df['body'].map(type) == np.ndarray,'body'] = np.nan # without mode set to null

#### Online Appraisal Flag
- 2% of online appraisal flag is missing
- Imputation: assume not online
- Note: inconsequential, not flagging imputation

# assume null is not online appraisal
df['online_appraisal_flag'].fillna(0, inplace=True)

#### Color
- 1% and <1% of color_appraisal and color are missing
- Imputation: replace with most common colors
- Note: inconsequential, so not flagging

# impute
df['color_appraisal'].fillna(df['color_appraisal'].mode()[0], inplace=True) # black
df['color'].fillna(df['color'].mode()[0], inplace=True) # white

#### MPGs, Cylinders, Engine Displacement, and Fuel Capacities
- <1% each
- Imputation: assume average
- Note: inconsequential, so not flagging

# impute
df['mpg_city'].fillna(df['mpg_city'].mean(), inplace=True)
df['mpg_highway'].fillna(df['mpg_highway'].mean(), inplace=True)
df['fuel_capacity_appraisal'].fillna(df['fuel_capacity_appraisal'].mean(), inplace=True)
df['fuel_capacity'].fillna(df['fuel_capacity'].mean(), inplace=True)
df['mpg_city_appraisal'].fillna(df['mpg_city_appraisal'].mean(), inplace=True)
df.loc[df['cylinders'] == 0,'cylinders'] = df['cylinders'].mean()
df.loc[df['engine_disp'] == 0,'engine_disp'] = df['engine_disp'].mean()

High correlation between city and highway mpg. Combining measure for simplicity

# correlation
print(df['mpg_city'].corr(df['mpg_highway']))
print(df['mpg_city_appraisal'].corr(df['mpg_highway_appraisal']))

# compute EPA combined measure
mpg = (0.55*df['mpg_city'] + 0.45*df['mpg_highway'])
mpg_appraisal = (0.55*df['mpg_city_appraisal'] + 0.45*df['mpg_highway_appraisal'])

df.insert(df.columns.get_loc('mpg_city'), 'mpg', mpg)
df.insert(df.columns.get_loc('mpg_city_appraisal'), 'mpg_appraisal', mpg_appraisal)
df.drop(columns=['mpg_city','mpg_highway','mpg_city_appraisal','mpg_highway_appraisal'], inplace=True)

### Confirm

# na counts
pd.DataFrame({'count':df.isna().sum()
             ,'percent':(df.isna().sum()/df.shape[0]).round(decimals=2)}
    ).sort_values(by='count', ascending=False)


# 0 counts in numeric vars
(df[['model_year','engine_disp','cylinders','mpg','horsepower','fuel_capacity'
    ,'model_year_appraisal','engine_disp_appraisal','cylinders_appraisal','mpg_appraisal','horsepower_appraisal','fuel_capacity_appraisal']] < 0.1).sum()

## Validation

# would think these counts would stop increasing as a make/model/trim has same features
print(len(df[['make','model']].value_counts(dropna=False)))
print(len(df[['make','model','premium_trim']].value_counts(dropna=False)))
print(len(df[['make','model','premium_trim','body']].value_counts(dropna=False)))
print(len(df[['make','model','premium_trim','body','engine_disp']].value_counts(dropna=False)))
print(len(df[['make','model','premium_trim','body','engine_disp','cylinders']].value_counts(dropna=False)))
print(len(df[['make','model','premium_trim','body','engine_disp','cylinders','horsepower']].value_counts(dropna=False)))
print(len(df[['make','model','premium_trim','body','engine_disp','cylinders','horsepower','fuel_capacity']].value_counts(dropna=False)))

### Engine
Here I've discovered a reliability problem with engine data
- Make/model/engine should determine cylinders/horsepower/fuel_capacity/mpg_*

print(df.columns)

pd.set_option('display.max_rows', 200)

test = df[df['model'] == 'HXQ_8']
test = test.groupby(['model_year','make','model','engine_disp','cylinders','horsepower'], dropna=False, as_index=False).size().sort_values(by = ['engine_disp','cylinders','horsepower'])
test = test[(test['size'] > 0) & (test['model_year'] == 2015)]
test['pct'] = (test['size']/test['size'].sum()).round(2)
test.sort_values(by = 'pct', ascending = False)

##### Ideas
- take the top-k engine configurations and make others confirm

## Write Out Data

file = 'data/cleaned_data.pkl'
df.to_pickle(file)
