==> 2_analysis.ipynb <==
# CarMax Analytics Showcase
- Competition: https://analyticsshowcase.carmax.com
- Data: https://github.com/kmx-analytics-showcase/Winter-2023

## Business Analysis
- In-depth analysis for business ideas

# environment details

from platform import python_version
print('python',python_version())

from importlib.metadata import version
print('numpy',version('numpy'))
print('pandas',version('pandas'))
print('pandas_profiling',version('pandas_profiling'))
print('pySankey',version('pySankey'))

import numpy as np
import pandas as pd

## Data Load

# load clean data
file = 'data/data_cleaned.csv'
df = pd.read_csv(file)
df.shape

print(df.columns)

## Market Differentiation

## Vehicle Segments

## Customer Preferences

### Brand Loyalty
