==> 3_model.ipynb <==
# CarMax Analytics Showcase
- Competition: https://analyticsshowcase.carmax.com
- Data: https://github.com/kmx-analytics-showcase/Winter-2023

## Recommendation System

import numpy as np
import pandas as pd

## Data Load

# load clean data
file = 'data/cleaned_data.pkl'
df = pd.read_pickle(file)
df.shape

## Modeling Thoughts
- too many makes/models/etc to be a classification problem

- need fewer features and targets
	- profile each model to come up with standard types (principal component analysis)

df.nunique()

## Feature Engineering
- We're starting with 13 features on a car
    - objective
        - price
    - condition
        - model_year
        - mileage
    - taste
        - make
        - model
        - premium_trim
        - body
        - color
    - kinetic features
        - engine_disp
        - cylinders
        - mpg
        - horsepower
        - fuel_capacity

## Price

## Model Year

## Mileage

## Make

## Model

## Premium Trim

## Body

## Color

## Kinetics
- Body x Kinetics
